make decompression api always provide an io.BufferedReader
Previously a BufferedReader was only provided when the data
was uncompressed, and otherwise a raw file object (the
read end of a compression command's pipe) was provided.  This
pipe was then rewrapped in a BufferedReader by the "unpack" API
to auto-detect the fileformat. If the datasource was uncompressed,
unpack would rewrap a BufferedReader twice and be unable to handle
the input.

Now the compression.decompressed() api always yields a BufferedReader
and this is assumed by the unpack code so there is no double wrapping
of input streams.

Fixes #90