use getpass.getuser() instead of relying on just $USER envvar
sandbox command previously defaulted to $USER envvar, but this
breaks in cases where that envvar doesn't exist.  This was observed
in at least some tox based py.test runs and could happen in real
world cases.  getpass is a bit more robust in this regard and I
think it is no less unsafe. This is mostly a stopgap measure until
this bit of functionality gets revisited in the near future.