Modify sandbox.sh start logic
Previously start waited for the mysqld socket file to appear as
the test for a successful service start.  Now start waits for the
pid-file to appear. This avoids a bug when starting a Galera enabled
mysqld, where mysqld may actually be started twice as part of the
recovery process, but startup fails after the first unix socket
appears.  Testing for pid-file is also a more standard practice
among mysql initscripts.

Resolves issue #88